+++ Source/ACE/Entity/Player.cs
@@ -198,10 +198,10 @@ namespace ACE.Entity
             set { character.TotalLogins = value; }
         }
 
-        public Player(Session session) : base(ObjectType.Creature, session.CharacterRequested.Guid)
+        public Player(Session session) : base(ObjectType.Creature, session.CharacterRequested.Guid, "Player", 1, ObjectDescriptionFlag.Stuck | ObjectDescriptionFlag.Player | ObjectDescriptionFlag.Attackable, WeenieHeaderFlag.ItemCapacity | WeenieHeaderFlag.ContainerCapacity | WeenieHeaderFlag.Usable | WeenieHeaderFlag.BlipColour | WeenieHeaderFlag.Radar, new Position(0, 0, 0, 0, 0, 0, 0, 0))
         {
             Session = session;
-            DescriptionFlags |= ObjectDescriptionFlag.Stuck | ObjectDescriptionFlag.Player | ObjectDescriptionFlag.Attackable;
+
             // This is the default send upon log in and the most common.   Anything with a velocity will need to add that flag.
             PositionFlag |= UpdatePositionFlag.ZeroQx | UpdatePositionFlag.ZeroQy | UpdatePositionFlag.Contact | UpdatePositionFlag.Placement;
             Name = session.CharacterRequested.Name;
@@ -214,7 +214,6 @@ namespace ACE.Entity
 
             SetPhysicsState(PhysicsState.IgnoreCollision | PhysicsState.Gravity | PhysicsState.Hidden | PhysicsState.EdgeSlide, false);
             PhysicsData.PhysicsDescriptionFlag = PhysicsDescriptionFlag.CSetup | PhysicsDescriptionFlag.MTable | PhysicsDescriptionFlag.Stable | PhysicsDescriptionFlag.Petable | PhysicsDescriptionFlag.Position;
-            WeenieFlags = WeenieHeaderFlag.ItemCapacity | WeenieHeaderFlag.ContainerCapacity | WeenieHeaderFlag.Usable | WeenieHeaderFlag.BlipColour | WeenieHeaderFlag.Radar;
 
             // apply defaults.  "Load" should be overwriting these with values specific to the character
             PhysicsData.MTableResourceId = 0x09000001u;
@@ -366,7 +365,7 @@ namespace ACE.Entity
                 // break if we reach max
                 if (character.Level == maxLevel.Level)
                 {
-                    PlayParticleEffect(Effect.WeddingBliss);
+                    PlayParticleEffect(Network.Enum.PlayScript.WeddingBliss);
                     break;
                 }
             }
@@ -391,7 +390,7 @@ namespace ACE.Entity
                 else
                     Session.Network.EnqueueSend(levelUp, levelUpMessage, xpUpdateMessage, currentCredits);
                 // play level up effect
-                PlayParticleEffect(Effect.LevelUp);
+                PlayParticleEffect(Network.Enum.PlayScript.LevelUp);
             }
         }
 
@@ -419,7 +418,7 @@ namespace ACE.Entity
                 if (IsAbilityMaxRank(ranks, isSecondary))
                 {
                     // fireworks
-                    PlayParticleEffect(Effect.WeddingBliss);
+                    PlayParticleEffect(Network.Enum.PlayScript.WeddingBliss);
                     messageText = $"Your base {ability} is now {newValue} and has reached its upper limit!";
                 }
                 else
@@ -567,7 +566,7 @@ namespace ACE.Entity
                 if (IsSkillMaxRank(ranks, status))
                 {
                     // fireworks on rank up is 0x8D
-                    PlayParticleEffect(Effect.WeddingBliss);
+                    PlayParticleEffect(Network.Enum.PlayScript.WeddingBliss);
                     messageText = $"Your base {skill} is now {newValue} and has reached its upper limit!";
                 }
                 else
@@ -584,9 +583,9 @@ namespace ACE.Entity
         }
 
         // plays particle effect like spell casting or bleed etc..
-        public void PlayParticleEffect(Effect effectId)
+        public void PlayParticleEffect(PlayScript effectId)
         {
-            var effectEvent = new GameMessageEffect(this.Guid, effectId);
+            var effectEvent = new GameMessageScript(this.Guid, effectId);
             Session.Network.EnqueueSend(effectEvent);
         }
 
@@ -935,7 +934,11 @@ namespace ACE.Entity
                         return;
 
                 if (!sendUpdate)
+                {
                     clientObjectList.Add(worldObject.Guid, WorldManager.PortalYearTicks);
+                    worldObject.PlayScript(this.Session);
+                }
+
                 else
                     clientObjectList[worldObject.Guid] = WorldManager.PortalYearTicks;
             }