+++ changelog.md
@@ -1,9 +1,18 @@
 # ACEmulator Change Log
+
 ### 2017-04-03
 [Jyrus]
 * Add changes to ACE-World ace_object and weenie tables to widen the indexing restrictions to allow more fields to be used for unique indexing
 * Add Portal weenie table data
 * Add Portal object table data
+
+[Mogwai]
+* Velocity vector fixed
+* Fixed protection level of weenie class in mutable world object.
+
+### 2017-04-02
+[StackOverflow]
+* SpellFactory + World Object Fixes and a real Vector3  - Improvements- AceVector
 
 ### 2017-03-31
 [fantoms]
@@ -13,6 +22,10 @@
 * Changed a log4net debug line.
 * Changed the name class name `CharacterPositionType` to `PositionType`
 * Reworked `Position.cs` with ORM.
+
+### 2017-03-30
+[StackOverflow]
+* Start of a SpellFactory + World Object Fixes and a real Vector3 - AceVector
 
 ### 2017-03-29
 [fantoms]
@@ -75,3 +88,4 @@
 ### 2017-03-25
 [Mogwai]
 * weenies weenies everywhere. object structure pass 1
+